+++ src/rule-definitions.ts
@@ -1,4 +1,5 @@
 import type { RuleDefinition } from '#src/common/rule.js';
+import banUnboundedStringType from '#src/rules/ban-unbounded-string-type.js';
 import fieldNameCamelCase from '#src/rules/field-name-camel-case.js';
 import fieldNameMappingSnakeCase from '#src/rules/field-name-mapping-snake-case.js';
 import fieldOrder from '#src/rules/field-order.js';
@@ -14,6 +15,7 @@ import requireFieldType from '#src/rules/require-field-type.js';
 import requireField from '#src/rules/require-field.js';
 
 export default [
+  banUnboundedStringType,
   fieldNameCamelCase,
   fieldNameMappingSnakeCase,
   fieldOrder,